chore: move code to create an extended JSON array of findings to utility class to share it
