+++ plugin-core/src/main/java/appland/webviews/findingDetails/FindingDetailsEditor.java
@@ -2,18 +2,17 @@ package appland.webviews.findingDetails;
 
 import appland.AppMapBundle;
 import appland.AppMapPlugin;
-import appland.webviews.appMap.AppMapFileEditor;
-import appland.webviews.appMap.AppMapFileEditorState;
-import appland.files.AppMapFiles;
 import appland.files.FileLocation;
-import appland.problemsView.ScannerProblem;
+import appland.problemsView.FindingsUtil;
+import appland.problemsView.ResolvedStackLocation;
 import appland.utils.GsonUtils;
 import appland.webviews.WebviewEditor;
+import appland.webviews.appMap.AppMapFileEditor;
+import appland.webviews.appMap.AppMapFileEditorState;
 import appland.webviews.findings.FindingsOverviewEditorProvider;
 import com.google.gson.*;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.ModalityState;
-import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.fileEditor.FileEditorManager;
 import com.intellij.openapi.fileEditor.OpenFileDescriptor;
 import com.intellij.openapi.project.Project;
@@ -26,16 +25,11 @@ import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.File;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 import static appland.utils.GsonUtils.singlePropertyObject;
 
 public class FindingDetailsEditor extends WebviewEditor<Void> {
-    private static final int STACK_TRACE_CHARACTER_LIMIT = 50;
 
     public FindingDetailsEditor(@NotNull Project project, @NotNull VirtualFile file) {
         super(project, file);
@@ -103,7 +97,7 @@ public class FindingDetailsEditor extends WebviewEditor<Void> {
         assert findings != null;
 
         payload.addProperty("page", "finding-details");
-        payload.add("data", singlePropertyObject("findings", createFindingsArray(findings)));
+        payload.add("data", singlePropertyObject("findings", FindingsUtil.createFindingsArray(gson, project, findings)));
     }
 
     private void openEditorForLocation(@NotNull ResolvedStackLocation location) {
@@ -140,118 +134,5 @@ public class FindingDetailsEditor extends WebviewEditor<Void> {
                         AppMapBundle.get("webview.findingDetails.appMapNotFound.title"));
             });
         }
-    }
-
-    private @NotNull JsonArray createFindingsArray(@NotNull List<ScannerProblem> findings) {
-        var findingsJSON = new JsonArray();
-        for (var finding : findings) {
-            findingsJSON.add(createFindingItemJson(finding));
-        }
-        return findingsJSON;
-    }
-
-    private @NotNull JsonObject createFindingItemJson(@NotNull ScannerProblem finding) {
-        var jsonItem = new JsonObject();
-        jsonItem.add("finding", gson.toJsonTree(finding.getFinding()));
-        jsonItem.add("appMapUri", createAppMapUriJson(finding));
-        jsonItem.add("problemLocation", createProblemLocationJson(finding));
-        jsonItem.add("stackLocations", ReadAction.compute(() -> createStackLocationsJson(finding)));
-        jsonItem.add("ruleInfo", createRuleInfoJson(finding));
-        jsonItem.addProperty("appMapName", findAppMapName(finding));
-        return jsonItem;
-    }
-
-    // follows VSCode's "filterFinding"
-    private @Nullable String findAppMapName(ScannerProblem finding) {
-        var appMapFile = AppMapFiles.findAppMapSourceFile(finding.getFindingsFile());
-        if (appMapFile == null) {
-            return null;
-        }
-
-        var filename = appMapFile.getName();
-        var index = filename.indexOf('.');
-        return index == -1 ? filename : filename.substring(0, index);
-    }
-
-    private @NotNull JsonElement createProblemLocationJson(@NotNull ScannerProblem finding) {
-        var location = finding.getFinding().getProblemLocation();
-        if (location == null) {
-            return JsonNull.INSTANCE;
-        }
-
-        var data = new JsonObject();
-        data.add("range", newRangeJson(location));
-        data.add("uri", singlePropertyObject("path", location.filePath));
-        return data;
-    }
-
-    private @NotNull JsonElement newRangeJson(FileLocation location) {
-        return location == null || location.line == null
-                ? JsonNull.INSTANCE
-                : singlePropertyObject("line", location.line);
-    }
-
-    private @NotNull JsonElement createStackLocationsJson(ScannerProblem finding) {
-        if (finding == null) {
-            return new JsonArray();
-        }
-
-        var baseFile = finding.getFindingsFile();
-        var stackLocations = finding.getFinding().stack.stream()
-                .map(frame -> resolveStackFrame(frame, baseFile))
-                .filter(Objects::nonNull)
-                .collect(Collectors.toUnmodifiableList());
-
-        var json = new JsonArray();
-        for (var location : stackLocations) {
-            json.add(gson.toJsonTree(location));
-        }
-        return json;
-    }
-
-    private @NotNull JsonElement createRuleInfoJson(@NotNull ScannerProblem finding) {
-        return gson.toJsonTree(finding.getFinding().ruleInfo);
-    }
-
-    // follows VSCode's "resolveAppMapUri"
-    private @NotNull JsonElement createAppMapUriJson(@NotNull ScannerProblem finding) {
-        var appMapFile = AppMapFiles.findAppMapSourceFile(finding.getFindingsFile());
-        if (appMapFile == null) {
-            return JsonNull.INSTANCE;
-        }
-
-        // adds the state of the webview as URI anchor
-        var state = AppMapFileEditorState
-                .createViewFlowState(finding.getFinding().getEventId(), finding.getFinding().relatedEvents)
-                .jsonState;
-        return singlePropertyObject("path", appMapFile.toNioPath() + "#" + state);
-    }
-
-    public static @NotNull String truncatePath(@NotNull String path, @NotNull Character separator) {
-        if (path.length() <= STACK_TRACE_CHARACTER_LIMIT) {
-            return path;
-        }
-
-        var separatorString = "" + separator;
-        while (path.contains(separatorString) && path.length() > STACK_TRACE_CHARACTER_LIMIT) {
-            path = path.substring(path.indexOf(separator) + 1);
-        }
-        return "..." + separator + path;
-    }
-
-    private @Nullable ResolvedStackLocation resolveStackFrame(@NotNull String frame, @NotNull VirtualFile baseFile) {
-        var location = FileLocation.parse(frame);
-        if (location == null) {
-            return null;
-        }
-
-        var resolvedFile = location.resolveFilePath(project, baseFile);
-        if (resolvedFile == null || !resolvedFile.isInLocalFileSystem()) {
-            return null;
-        }
-
-        var nativePath = resolvedFile.toNioPath().toString();
-        var truncatedPath = truncatePath(nativePath, File.separatorChar);
-        return new ResolvedStackLocation(nativePath, truncatedPath + location.getSuffix(), location.getZeroBasedLine(0));
     }
 }
